web_app/pyproject.toml
[tool.poetry]
name = "web-app"
version = "0.1.0"
description = ""
authors = ["LevytskyiS <levytskyi.sv@gmail.com>"]
readme = "README.md"
packages = [{include = "web_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
django-environ = "^0.10.0"
django-mptt = "^0.14.0"
pillow = "^10.0.0"
django-ckeditor = "^6.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
